requirements.txt
pandas
fastapi
uvicorn
python-dotenv
httpx
openai